# packages/spanish-api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Spanish Learning API",
    "version": "1.0.0",
    "description": "API for learning Spanish vocabulary"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
      "Word": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "word": {
            "type": "string"
          },
          "translation": {
            "type": "string"
          },
          "example": {
            "type": "string",
            "nullable": true
          },
          "exampleTranslation": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "goodAnswers": {
            "type": "integer"
          },
          "goodAnswersStreak": {
            "type": "integer"
          },
          "badAnswers": {
            "type": "integer"
          },
          "lastAnswerTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "categoryId": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "WordStats": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "word": {
            "type": "string"
          },
          "translation": {
            "type": "string"
          },
          "goodAnswers": {
            "type": "integer"
          },
          "badAnswers": {
            "type": "integer"
          },
          "lastAnswerTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "totalAnswers": {
            "type": "integer"
          },
          "successRate": {
            "type": "number"
          }
        }
      }
    }
  },
  "paths": {
    "/ai/generate-examples": {
      "post": {
        "summary": "Generate example sentences for given words",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "words": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "targetLang": {
                    "type": "string",
                    "default": "Spanish"
                  },
                  "sentencesPerWord": {
                    "type": "number",
                    "default": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Example sentences generated successfully"
          },
          "400": {
            "description": "Invalid request - words array is required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/ai/generate-related-words": {
      "post": {
        "summary": "Generate related words for a given word",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "word": {
                    "type": "string"
                  },
                  "targetLang": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Related words retrieved successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/ai/generate-memory-associations": {
      "post": {
        "summary": "Generate memory association tips for a word",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "word": {
                    "type": "string"
                  },
                  "targetLang": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Memory tips retrieved successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/words": {
      "get": {
        "responses": {
          "200": {
            "description": "List of words",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Word"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/words/not-learned": {
      "get": {
        "summary": "Get not learned words",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of words to return"
          }
        ],
        "responses": {
          "200": {
            "description": "List of not learned words"
          }
        }
      }
    },
    "/words/not-learned-with-examples": {
      "get": {
        "summary": "Get not learned words with example sentences",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of words to return"
          }
        ],
        "responses": {
          "200": {
            "description": "List of not learned words with examples",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Word"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "example": {
                            "type": "object",
                            "properties": {
                              "original": {
                                "type": "string"
                              },
                              "english": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/words/answer": {
      "post": {
        "summary": "Save answer for a word",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "wordId": {
                    "type": "integer"
                  },
                  "isCorrect": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Answer saved successfully"
          }
        }
      }
    },
    "/words/{wordId}/stats": {
      "get": {
        "summary": "Get statistics for a word",
        "parameters": [
          {
            "in": "path",
            "name": "wordId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the word"
          }
        ],
        "responses": {
          "200": {
            "description": "Word statistics"
          },
          "404": {
            "description": "Word not found"
          }
        }
      }
    },
    "/words/{wordId}/skip": {
      "post": {
        "summary": "Skip a word",
        "parameters": [
          {
            "in": "path",
            "name": "wordId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the word to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Word skipped successfully"
          },
          "404": {
            "description": "Word not found"
          }
        }
      }
    },
    "/words/{wordId}/streak": {
      "post": {
        "summary": "Force set streak for a word",
        "parameters": [
          {
            "in": "path",
            "name": "wordId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the word"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "streak": {
                    "type": "integer",
                    "description": "Streak to set"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/words/learning-stats": {
      "get": {
        "summary": "Get learning statistics",
        "responses": {
          "200": {
            "description": "Learning statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalWords": {
                      "type": "integer",
                      "description": "Total number of words in the system"
                    },
                    "learnedWords": {
                      "type": "integer",
                      "description": "Number of words marked as learned"
                    },
                    "skippedWords": {
                      "type": "integer",
                      "description": "Number of words marked as skipped"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch learning stats"
          }
        }
      }
    },
    "/words/random-not-learned-with-examples": {
      "get": {
        "summary": "Get random not learned words with example sentences",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of words to return"
          },
          {
            "in": "query",
            "name": "poolSize",
            "schema": {
              "type": "integer"
            },
            "description": "Size of the pool to randomly select from"
          }
        ],
        "responses": {
          "200": {
            "description": "List of random not learned words with examples",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Word"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "example": {
                            "type": "object",
                            "properties": {
                              "original": {
                                "type": "string"
                              },
                              "english": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}