# packages/spanish-web/src/api/generated/spanishLearningAPI.schemas.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Spanish Learning API
 * API for learning Spanish vocabulary
 * OpenAPI spec version: 1.0.0
 */
export type GetWordsRandomNotLearnedWithExamples200ItemAllOfExample = {
  english?: string;
  original?: string;
};

export type GetWordsRandomNotLearnedWithExamples200ItemAllOf = {
  example?: GetWordsRandomNotLearnedWithExamples200ItemAllOfExample;
};

export type GetWordsRandomNotLearnedWithExamples200Item = Word & GetWordsRandomNotLearnedWithExamples200ItemAllOf;

export type GetWordsRandomNotLearnedWithExamplesParams = {
/**
 * Number of words to return
 */
limit?: number;
/**
 * Size of the pool to randomly select from
 */
poolSize?: number;
};

export type GetWordsLearningStats200 = {
  /** Number of words marked as learned */
  learnedWords?: number;
  /** Number of words marked as skipped */
  skippedWords?: number;
  /** Total number of words in the system */
  totalWords?: number;
};

export type PostWordsWordIdStreakBody = {
  /** Streak to set */
  streak?: number;
};

export type PostWordsAnswerBody = {
  isCorrect?: boolean;
  wordId?: number;
};

export type GetWordsNotLearnedWithExamples200ItemAllOfExample = {
  english?: string;
  original?: string;
};

export type GetWordsNotLearnedWithExamples200ItemAllOf = {
  example?: GetWordsNotLearnedWithExamples200ItemAllOfExample;
};

export type GetWordsNotLearnedWithExamples200Item = Word & GetWordsNotLearnedWithExamples200ItemAllOf;

export type GetWordsNotLearnedWithExamplesParams = {
/**
 * Number of words to return
 */
limit?: number;
};

export type GetWordsNotLearnedParams = {
/**
 * Number of words to return
 */
limit?: number;
};

export type PostAiGenerateMemoryAssociationsBody = {
  targetLang?: string;
  word?: string;
};

export type PostAiGenerateRelatedWordsBody = {
  targetLang?: string;
  word?: string;
};

export type PostAiGenerateExamplesBody = {
  sentencesPerWord?: number;
  targetLang?: string;
  words?: string[];
};

export interface WordStats {
  badAnswers?: number;
  goodAnswers?: number;
  id?: number;
  /** @nullable */
  lastAnswerTime?: string | null;
  successRate?: number;
  totalAnswers?: number;
  translation?: string;
  word?: string;
}

export interface Word {
  badAnswers?: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  example?: string | null;
  /** @nullable */
  exampleTranslation?: string | null;
  goodAnswers?: number;
  goodAnswersStreak?: number;
  id?: number;
  /** @nullable */
  lastAnswerTime?: string | null;
  /** @nullable */
  notes?: string | null;
  translation?: string;
  word?: string;
}